+++ lib/surveyor/models/dependency_condition_methods.rb
@@ -37,8 +37,8 @@ module Surveyor
       def to_hash(response_set)
         # all responses to associated question
         responses = question.blank? ? [] : response_set.responses.where("responses.answer_id in (?)", question.answer_ids).all
-        if self.operator.match /^count(>|>=|<|<=|=|!=)\d+$/
-          op, i = self.operator.scan(/^count(>|>=|<|<=|=|!=)(\d+)$/).flatten
+        if self.operator.match /^count(>|>=|<|<=|==|!=)\d+$/
+          op, i = self.operator.scan(/^count(>|>=|<|<=|==|!=)(\d+)$/).flatten
           # logger.warn({rule_key.to_sym => responses.count.send(op, i.to_i)})
           return {rule_key.to_sym => (op == "!=" ? !responses.count.send("==", i.to_i) : responses.count.send(op, i.to_i))}
         elsif operator == "!=" and (responses.blank? or responses.none?{|r| r.answer.id == self.answer.id})