+++ CHANGELOG.md
@@ -4,6 +4,13 @@ History for Surveyor
 1.0.2
 -----
 
+### Fixes
+
+- fixing and documenting count operators on dependency conditions
+
+### Infrastructure
+
+- basic spec for the surveyor task
 
 1.0.1
 ------
@@ -14,12 +21,15 @@ History for Surveyor
   now exclude the question or question group from the DOM. These display types are
   used to inject data (responses) into surveys. Note, custom_class => "hidden" doesn't
   have any effect until a custom css rule is created by the end user. (#197)
+
 - more readable parser and more strict parser method aliases (#278)
 
 ### Fixes
 
 - Replaced deprecated ActiveRecord::Errors#each_full with ActiveRecord::Errors#full_messages. (#363)
+
 - fixing dependency condition evaluation where #Response.*_value is nil. (#297)
+
 - fixing grid answers leak, introduced in 5baa7ac3. thanks @jsurrett (#375, #377)
 
 1.0.0